Separate simulation logic from Analyze section
